# shared/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WMS API",
    "description": "Wave 5 增值模块与横向能力契约",
    "license": {
      "name": ""
    },
    "version": "0.0.5-wave-5-value-added"
  },
  "paths": {
...
    "/api/v1/system-dictionaries/{dict_code}/items": {
      "get": {
        "tags": [
          "system-dictionary"
        ],
        "operationId": "list_system_dictionary_items",
        "parameters": [
          {
            "name": "dict_code",
            "in": "path",
            "description": "字典分类编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "effective_at",
            "in": "query",
            "description": "按指定时间查询有效字典项",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "按货主合并后的有效字典项",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDictionaryItemListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "字典分类不存在或停用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "运行时字典参数无效，fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system-dictionaries/{dict_code}/items/{item_code}": {
      "put": {
        "tags": [
          "system-dictionary"
        ],
        "operationId": "upsert_system_dictionary_item",
        "parameters": [
          {
            "name": "dict_code",
            "in": "path",
            "description": "字典分类编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_code",
            "in": "path",
            "description": "字典项编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "客户端生成的幂等键",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertSystemDictionaryItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "创建或更新字典项",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDictionaryItem"
                }
              }
            }
          },
          "400": {
            "description": "缺少或非法幂等键",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "字典分类不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "字典项参数或作用域非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/system-dictionaries/{dict_code}/items/{item_code}/disable": {
      "patch": {
        "tags": [
          "system-dictionary"
        ],
        "operationId": "disable_system_dictionary_item",
        "parameters": [
          {
            "name": "dict_code",
            "in": "path",
            "description": "字典分类编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item_code",
            "in": "path",
            "description": "字典项编码",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "客户端生成的幂等键",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableSystemDictionaryItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "停用字典项",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemDictionaryItem"
                }
              }
            }
          },
          "400": {
            "description": "缺少或非法幂等键",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "字典分类或字典项不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "字典项作用域非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DisableSystemDictionaryItemRequest": {
        "type": "object",
        "properties": {
          "disabled_reason": {
            "type": "string",
            "nullable": true
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "统一错误响应。",
        "required": [
          "code",
          "message",
          "severity",
          "details",
          "trace_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "业务错误码。"
          },
          "details": {
            "type": "object",
            "description": "关联详情。",
            "additionalProperties": true
          },
          "message": {
            "type": "string",
            "description": "中文错误消息。"
          },
          "retry_hint": {
            "type": "string",
            "description": "重试提示。",
            "nullable": true
          },
          "severity": {
            "type": "string",
            "description": "严重度。"
          },
          "trace_id": {
            "type": "string",
            "description": "链路追踪 ID。"
          }
        }
      },
...
      "PageMeta": {
        "type": "object",
        "description": "分页信息。",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32",
            "description": "本页数量。",
            "minimum": 0
          },
          "next_cursor": {
            "type": "string",
            "description": "下一页游标；为空表示无更多数据。",
            "nullable": true
          }
        }
      },
...
      "SystemDictionaryItem": {
        "type": "object",
        "description": "系统字典项。",
        "required": [
          "id",
          "dict_code",
          "item_code",
          "item_name",
          "enabled",
          "params",
          "source",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "dict_code": {
            "type": "string"
          },
          "disabled_reason": {
            "type": "string",
            "nullable": true
          },
          "effective_from": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "effective_to": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_code": {
            "type": "string"
          },
          "item_name": {
            "type": "string"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "params": {
            "type": "object",
            "description": "自由结构 JSON 对象。",
            "additionalProperties": true
          },
          "source": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SystemDictionaryItemListResponse": {
        "type": "object",
        "required": [
          "data",
          "page"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SystemDictionaryItem"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        }
      },
...
      "UpsertSystemDictionaryItemRequest": {
        "type": "object",
        "required": [
          "item_name",
          "enabled",
          "params"
        ],
        "properties": {
          "effective_from": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "effective_to": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "item_name": {
            "type": "string"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "params": {
            "type": "object",
            "description": "自由结构 JSON 对象。",
            "additionalProperties": true
          }
        }
      },
...
    }
  },
...
}